move duration mod to file
